GeniAnalysis/requirements.txt
acme==1.1.0
aioredis==1.3.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==19.3.0
autobahn==21.11.1
Automat==0.8.0
backports.zoneinfo==0.2.1
blinker==1.4
certbot==0.40.0
certbot-nginx==0.40.0
certifi==2019.11.28
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
chardet==3.0.4
charset-normalizer==2.0.11
Click==7.0
cloud-init==21.4
colorama==0.4.3
command-not-found==0.3
ConfigArgParse==0.13.0
configobj==5.0.6
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.1
django-crontab==0.7.1
entrypoints==0.3
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
hiredis==2.0.0
httplib2==0.14.0
httptools==0.3.0
hyperlink==21.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
iotop==0.6
Jinja2==2.10.1
josepy==1.2.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lz4==3.1.3
MarkupSafe==1.1.0
mock==3.0.5
more-itertools==4.2.0
msgpack==1.0.3
mysqlclient==1.4.4
netifaces==0.10.4
numpy==1.22.1
oauthlib==3.1.0
pandas==1.4.0
parsedatetime==2.4
pbr==5.4.5
pexpect==4.6.0
psycopg2-binary==2.9.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
PyGObject==3.36.0
PyHamcrest==1.9.0
PyICU==2.4.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.3.1
requests==2.27.1
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
requests-unixsocket==0.2.0
schedule==1.1.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.1
sqlparse==0.4.2
ssh-import-id==5.10
supervisor==4.1.0
systemd-python==234
timeloop==1.0.2
truedata-ws==4.2.4
Twisted==18.9.0
txaio==21.2.1
ubuntu-advantage-tools==27.5
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
uvicorn==0.17.1
uvloop==0.16.0
wadllib==1.3.3
watchgod==0.7
websocket==0.2.1
websocket-client==1.2.3
websockets==10.1
zipp==1.0.0
zope.component==4.3.0
zope.event==4.5.0
zope.hookable==5.0.0
zope.interface==4.7.1
